Relocate the converter stencils
